# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Versioned commerce API for the storefront and operations console
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/discounts/validate:
    post:
      operationId: validateDiscount
      tags: [storefront]
      summary: Validate a dynamic discount code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscountValidationInput"
      responses:
        "200":
          description: Discount validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscountValidation"
...
  /v1/checkout/create-order:
    post:
      operationId: createCheckoutOrder
      tags: [storefront]
      summary: Create a checkout order (owner derived from the bearer token, not the body)
      security:
        - customerSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [idempotencyKey, items, shippingAddress]
              properties:
                idempotencyKey: { type: string }
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      productId: { type: string }
                      quantity: { type: number }
                discountCode: { type: string }
                customerEmail: { type: string }
                customerName: { type: string }
                customerMobile: { type: string }
                shippingAddress: { type: string }
                deliveryLatitude: { type: number }
                deliveryLongitude: { type: number }
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                type: object
                properties:
                  orderId: { type: string }
                  razorpayOrderId: { type: string }
                  amountCents: { type: number }
                  currency: { type: string }
                  keyId: { type: string }
                  status: { type: string }
                  idempotencyKey: { type: string }
  /v1/checkout/verify-payment:
    post:
      operationId: verifyPayment
      tags: [storefront]
      summary: Verify a Razorpay payment signature and confirm the order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [idempotencyKey, razorpayOrderId, razorpayPaymentId]
              properties:
                idempotencyKey: { type: string }
                razorpayOrderId: { type: string }
                razorpayPaymentId: { type: string }
                razorpaySignature: { type: string }
      responses:
        "200":
          description: Payment verified
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  orderId: { type: string }
                  status: { type: string }
  /v1/checkout/orders:
    get:
      operationId: listMyOrders
      tags: [storefront]
      summary: List the signed-in customer's own orders (scoped to the bearer token)
      security:
        - customerSession: []
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: duration, in: query, schema: { type: string } }
        - { name: startDate, in: query, schema: { type: string } }
        - { name: endDate, in: query, schema: { type: string } }
        - { name: search, in: query, schema: { type: string } }
      responses:
        "200":
          description: Order history
          content:
            application/json:
              schema:
                type: array
                items: { type: object, additionalProperties: true }
  /v1/checkout/orders/{id}/cancel:
    post:
      operationId: cancelMyOrder
      tags: [storefront]
      summary: Cancel one of the signed-in customer's own pending orders
      security:
        - customerSession: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema: { type: object, additionalProperties: true }

  # ─── Customer authentication (bearer-token sessions) ─────────────────────
...
  /v1/admin/shop-settings:
    get:
      operationId: getShopSettings
      tags: [admin]
      summary: Read shop settings
      security:
        - clerkSession: []
      responses:
        "200":
          description: Shop settings
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
    put:
      operationId: updateShopSettings
      tags: [admin]
      summary: Update shop settings
      security:
        - clerkSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { type: object, additionalProperties: true }

  # ─── Staff RBAC (Clerk-protected; login is open) ─────────────────────────
...
components:
  securitySchemes:
    clerkSession:
      type: http
      scheme: bearer
      bearerFormat: Clerk
    customerSession:
      type: http
      scheme: bearer
      bearerFormat: Opaque
    staffSession:
      type: http
      scheme: bearer
      bearerFormat: Opaque
...
  schemas:
...
    DiscountValidationInput:
      type: object
      required: [code, subtotalCents, isFirstOrder]
      properties:
        code: { type: string }
        subtotalCents: { type: number, minimum: 0 }
        isFirstOrder: { type: boolean }
    DiscountValidation:
      type: object
      required: [valid, code, discountCents, message]
      properties:
        valid: { type: boolean }
        code: { type: string }
        discountCents: { type: number, minimum: 0 }
        message: { type: string }
...